pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel",
    "Cython",
    "numpy"
]
build-backend = "setuptools.build_meta"

[project]
name = "vEcoli"
version = "1.1.0"
description = "Whole Cell E. coli Model"
requires-python = "==3.12.8"
dependencies = [
    "numpy",
    "scipy",
    "biopython",
    "cvxpy",
    "cython",
    "dill",
    "duckdb",
    "ete3",
    "gcsfs",
    "pyarrow",
    "ipdb",
    "ipython",
    "jupyter",
    "line-profiler",
    "altair",
    "numba",
    "orjson",
    # Constraint can be removed once cvxpy supports newer ortools
    "ortools<=9.9.3963",
    "pandas",
    "polars",
    "pymunk",
    "scikit-image",
    "scikit-learn",
    "seaborn",
    "swiglpk",
    "sympy",
    "tqdm",
    "unum",
    "vivarium-core",
    "pysal",
    "opencv-python-headless",
    "statsmodels",
    "ecos",
    "nbclassic",
    "matplotlib",
    "pyqt5",
    "iteround",
    "wheel",
    "stochastic-arrow",
    "autograd",
    "typer",
    "astor",
    "process-bigraph>=0.0.30",
    "vivarium-interface>=0.0.4",
]

[project.scripts]
gateway = "api.gateway.main:cli"
crosspiler = "pbg.app:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "pre-commit",
]
docs = [
    "Sphinx",
    "sphinx-rtd-theme",
    "nbsphinx"
]
shared = [
    "cryptography",
]
api = [
    "broadcaster>=0.3.1",
    "cryptography>=44.0.2",
    "fastapi>=0.115.12",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "typer>=0.15.2",
    "uvicorn>=0.34.1",
    "websockets>=15.0.1",
]

[tool.setuptools]
packages = [
    # original packages
    "ecoli",
    "reconstruction",
    "runscripts",
    "validation",
    "wholecell",
    # packages added by Alex
    "pbg",  # TODO: eventually extract needed from this and remove
    "api",
    "migration"
]
